Add toggle sidebars buttons
